# openapi.yaml
openapi: 3.0.1
info:
  title: "AuleWeb Services"
  description: |-
    Questo sistema è volto alla gestione delle aule presenti in un qualsiasi Ateneo.
    //DA ESTENDERE
  contact:
    email: sviluppoWebAvanzato@student.univaq.it
  version: 1.0.0

servers:
  - url: 'http://localhost:8080/AuleWeb'

tags:
  - name: Autenticazione
  - name: Aule

paths:
  '/rest/auth/login':
    post:
      summary: Autenticazione Amministratore
      description: Verifica le credenziali di un amministratore che tenta di autenticarsi nel sistema
      operationId: doLogin #1
      tags:
        - Autenticazione
      requestBody:
        content:
          application/x-www-form-urlencoded: # vedere nel codice se usare json o form
            schema:
              $ref: '#/components/schemas/Credenziali'
        required: true
      responses:
        '200':
          description: Successful authentication
          content:
            text/plain:
              schema:
                type: string
                pattern: 'Bearer [a-z0-9A-Z-]+'
           
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        
        '500':
          $ref: '#/components/responses/GeneralError'

  '/rest/auth/logout':
    delete:
      summary: Deautenticazione Amministratore
      description: Consente il logout dell'amministratore
      operationId: doLogout #1
      tags:
        - Autenticazione
      security:
        - bearerAuth: []
      responses:
        '200':
          description: successfull logout
        
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        
        '500':
          $ref: '#/components/responses/GeneralError'
  
  '/rest/aule/csv':
    get:
      operationId: exportAuleCsv #2
      tags: 
        - Aule
      responses:
        '200':
          description: esportazione aule in CSV
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '500':
          $ref: '#/components/responses/GeneralError'
        '404':
          $ref: '#/components/responses/NotFound'
    
    post:
      operationId: importAuleCsv #2
      tags: 
        - Aule
      requestBody:
        required: true # vedere se bisogna mettere required anche agli altri requestbody delle altre post!!!
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      
      responses:
        '201':
          description: created
          headers:
            location:
              schema:
                type: string
                format: uri
          content:
            application/json: # vedere se effettivamente è un json!!!
              schema:
                type: string
                format: uri
        '500':
          $ref: '#/components/responses/GeneralError'
  
  
  '/rest/aule':
    post:
      operationId: addAula #3
      tags:
        - Aule
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Aula'
      responses:
        '201':
          description: created
          headers:
            location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                type: string
                format: uri
        
        '500':
          $ref: '#/components/responses/GeneralError'
  
  '/rest/aule/{id}/assignGruppo':
    post:
      operationId: assignGruppo #4
      tags:
        - Aule
      parameters: 
        - $ref: '#/components/parameters/idAula'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id: 
                  type: integer
      responses:
        '201':
          description: created
          headers:
            location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                type: string
                format: uri
        '500':
          $ref: '#/components/responses/GeneralError'
          
    get:
      operationId: getInfoAula #5
      tags:
        - Aule
      parameters: 
        - $ref: '#/components/parameters/idAula'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aula'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/GeneralError'
  
  '/rest/aule/{nome}/attrezzature':  
    get:
      operationId: getAttrezzatureAula #6
      tags: 
        - Aule
      parameters:
        - $ref: '#/components/parameters/idAula'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attrezzature'
        
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/GeneralError'
    
  '/rest/eventi':
    post:
      tags: 
        - Eventi
      operationId: addEvento #7
      requestBody: 
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Evento'
      responses:
        '201':
          description: created
          headers:
            location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                type: string
                format: uri
        
        '500':
          $ref: '#/components/responses/GeneralError'
    
    get:
      operationId: getEventiForRange #12
      tags:
        - Eventi
      parameters: 
        - $ref: '#/components/parameters/rangeStart'
        - $ref: '#/components/parameters/rangeEnd'
      responses:
        '200':
          description: OK
          content:
            text/calendar:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
      
          
  '/rest/eventi/{evento}':
    patch:
      operationId: updateEvento #8
      tags:
        - Eventi
      parameters:
        - $ref: '#/components/parameters/idEvento'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evento'
      responses:
        # capire se bisogna mettere pure 201 come possibile risposta!!!
        '204':
          description: no content
        '500':
          $ref: '#/components/responses/GeneralError'
        '404':
          $ref: '#/components/responses/NotFound'
    
    get:
      operationId: getInfoEvento #9
      tags:
        - Eventi
      parameters:
        - $ref: '#/components/parameters/idEvento'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evento'
        '500':
          $ref: '#/components/responses/GeneralError'
        '404':
          $ref: '#/components/responses/NotFound'
          
  '/rest/eventi/aule/{nome}': # vedere come filtrare la settimana. In base alla i-esima settimana???
  # nella URL ci va pure eventi?
    get:
      operationId: getEventiAulaSettimana #10
      tags:
        - Eventi
        - Aule
      parameters:
        - $ref: '#/components/parameters/idAula'
        - $ref: '#/components/parameters/numeroSettimana'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Eventi'
        '500':
          $ref: '#/components/responses/GeneralError'
        '404':
          $ref: '#/components/responses/NotFound'
  
  '/rest/eventi/attuali': # lista degli eventi attuali (non ancora terminati)
    get:
      operationId: getEventiAttuali #11
      tags:
      - Eventi
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Eventi'
        '500':
          $ref: '#/components/responses/GeneralError'
        '404':
          $ref: '#/components/responses/NotFound'

  # ritorna la lista dei prossimi eventi nelle prossime x ore (se non specificato, quelli delle prossime 3 ore)
  '/rest/eventi/prossimi':
    get:
      operationId: getEventiProssimi #11
      tags:
        - Eventi
      parameters: 
        - $ref: '#/components/parameters/oreProssime'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Eventi'
        '500':
          $ref: '#/components/responses/GeneralError'
        '404':
          $ref: '#/components/responses/NotFound'


components:
  schemas:

    Eventi:
      type: array
      items:
        $ref: '#/components/schemas/Evento'
        
    Evento:
      type: object
      properties:
        orario_inizio:
          type: string
          format: date-time
        orario_fine:
          type: string
          format: date-time
        descrizione:
          type: string
        nome_organizzatore:
          type: string
        email_responsabile:
          type: string
          format: email
        tipologia:
          type: string
          enum:
            - lezione
            - esame
            - seminario
            - parziale
            - riunione
            - lauree
            - altro
        id_master:
          type: integer
          format: int32
        id_aula:
          $ref: '#/components/schemas/Aula' # mettere url da richiamare in GET o il JSON object
        id_corso:
          $ref: '#/components/schemas/Corso'
        
    Corso:
      type: object
      properties:
        nome:
          type: string
    
    Attrezzature:
      type: array
      items:
        type: object
        properties:
          tipo:
            type: string
  
    Credenziali:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    
    Aula:
      type: object
      properties:
        nome:
          type: string
        luogo:
          type: string
        edificio:
          type: string
        piano:
          type: string
        capienza:
          type: integer
          format: int32
          default: 0
        emailResponsabile:
          type: string
        preseElettriche:
          type: integer
          format: int32
          default: 0
        preseRete:
          type: integer
          format: int32
          default: 0
        note:
          type: string
  
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: internalToken
      
  parameters:
    idAula:
      name: id
      in: path
      required: true
      schema:
        type: integer
        
    idEvento:
      name: evento
      in: path
      required: true
      schema:
        type: integer
        
    numeroSettimana:
      name: settimana
      in: query
      required: true
      schema:
        type: integer
    
    rangeStart:
      in: query
      name: rangeStart
      required: true
      schema:
        type: string
    
    rangeEnd:
      in: query
      name: rangeEnd
      required: true
      schema:
        type: string

    oreProssime :
      in: query
      name: oreProssime
      required: false
      schema:
        type: integer
        default: 3
  
  responses:
    GeneralError:
      description: errore generale
      content:
        text/plain:
          schema:
            type: string
            example: "Messaggio di errore"
    
    UnauthorizedError:
      description: bearer token mancante o non valido
    
    NotFound:
      description: elemento non trovato
      content: 
        text/plain:
          schema:
            type: string
            example: "Elemento insesistente"